feat (paging-setup) : add basic setup for paging3

+++ shared/src/commonMain/kotlin/com/example/kast/data/model/TmdbMovieResult.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 data class TmdbMovieResult(
 
     @SerialName("page")
-    val page: Long?,
+    val page: Int?,
 
     @SerialName("results")
     val results: List<TmdbMovie>?,
